argoslabs/data/pdf2txt/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# ...
PyPDF2